+++ agent/src/explainPollyError.ts
@@ -1,0 +1,18 @@
+export const explainPollyError = `
+                console.error(error)
+
+    ===================================================[ NOTICE ]=======================================================
+    If you get PollyError or unexpected diff, you might need to update recordings to match your changes.
+    Run the following commands locally to update the recordings:
+
+      source agent/scripts/export-cody-http-recording-tokens.sh
+      pnpm update-agent-recordings
+      # Press 'u' to update the snapshots if the new behavior makes sense. It's
+      # normal that the LLM returns minor changes to the wording.
+      git commit -am "Update agent recordings"
+
+
+    More details in https://github.com/sourcegraph/cody/tree/main/agent#updating-the-polly-http-recordings
+    ====================================================================================================================
+
+    `
