+++ agent/src/index.test.ts
@@ -19,32 +19,9 @@ import { TESTING_CREDENTIALS } from '../../vscode/src/testutils/testing-credenti
 import { TestClient, asTranscriptMessage } from './TestClient'
 import { TestWorkspace } from './TestWorkspace'
 import { decodeURIs } from './decodeURIs'
-import type {
-    CustomChatCommandResult,
-    CustomEditCommandResult,
-    EditTask,
-    Requests,
-} from './protocol-alias'
+import { explainPollyError } from './explainPollyError'
+import type { Requests } from './protocol-alias'
 import { trimEndOfLine } from './trimEndOfLine'
-
-const explainPollyError = `
-                console.error(error)
-
-    ===================================================[ NOTICE ]=======================================================
-    If you get PollyError or unexpected diff, you might need to update recordings to match your changes.
-    Run the following commands locally to update the recordings:
-
-      source agent/scripts/export-cody-http-recording-tokens.sh
-      pnpm update-agent-recordings
-      # Press 'u' to update the snapshots if the new behavior makes sense. It's
-      # normal that the LLM returns minor changes to the wording.
-      git commit -am "Update agent recordings"
-
-
-    More details in https://github.com/sourcegraph/cody/tree/main/agent#updating-the-polly-http-recordings
-    ====================================================================================================================
-
-    `
 
 const workspace = new TestWorkspace(path.join(__dirname, '__tests__', 'example-ts'))
 
@@ -893,10 +870,7 @@ describe('Agent', () => {
                   }
 
                   return a - b
-              }
-
-
-              "
+              }"
             `,
                 explainPollyError
             )
@@ -933,8 +907,7 @@ describe('Agent', () => {
                       expect(trickyLogic(-5, 3)).toBe(-8);
                       expect(trickyLogic(5, -3)).toBe(8);
                   });
-              });
-              "
+              });"
             `,
                 explainPollyError
             )
@@ -951,10 +924,9 @@ describe('Agent', () => {
                 obtained =>
                     expect(obtained).toMatchInlineSnapshot(
                         `
-                    "export function sum(c: number, b: number): number {
-                        /* CURSOR */
-                    }
-                    "
+                      "export function sum(c: number, b: number): number {
+                          /* CURSOR */
+                      }"
                     `,
                         explainPollyError
                     )
@@ -1002,8 +974,7 @@ describe('Agent', () => {
                   			</ul>
                   		</>
                   	);
-                  }
-                  "
+                  }"
                 `,
                     explainPollyError
                 )
@@ -1030,8 +1001,7 @@ describe('Agent', () => {
                   export function Heading({ text, level = 1 }: HeadingProps) {
                       const HeadingTag = \`h\${level}\` as keyof JSX.IntrinsicElements;
                       return <HeadingTag>{text}</HeadingTag>;
-                  }
-                  "
+                  }"
                 `,
                     explainPollyError
                 )
@@ -1050,8 +1020,7 @@ describe('Agent', () => {
                    */
                   export function sum(a: number, b: number): number {
                       /* CURSOR */
-                  }
-                  "
+                  }"
                 `,
                     explainPollyError
                 )
@@ -1064,22 +1033,21 @@ describe('Agent', () => {
                 obtained =>
                     expect(obtained).toMatchInlineSnapshot(
                         `
-                  "const foo = 42
+                      "const foo = 42
 
-                  export class TestClass {
-                      constructor(private shouldGreet: boolean) {}
+                      export class TestClass {
+                          constructor(private shouldGreet: boolean) {}
 
-                          /**
-                       * Logs "Hello World!" to the console if \`shouldGreet\` is true.
-                       */
-                  public functionName() {
-                          if (this.shouldGreet) {
-                              console.log(/* CURSOR */ 'Hello World!')
+                              /**
+                           * Logs "Hello World!" to the console if \`shouldGreet\` is true.
+                           */
+                      public functionName() {
+                              if (this.shouldGreet) {
+                                  console.log(/* CURSOR */ 'Hello World!')
+                              }
                           }
-                      }
-                  }
-                  "
-                `,
+                      }"
+                    `,
                         explainPollyError
                     )
             )
@@ -1090,23 +1058,22 @@ describe('Agent', () => {
                 'TestLogger.ts',
                 obtained =>
                     expect(obtained).toMatchInlineSnapshot(`
-                  "const foo = 42
-                  export const TestLogger = {
-                      startLogging: () => {
-                          // Do some stuff
+                      "const foo = 42
+                      export const TestLogger = {
+                          startLogging: () => {
+                              // Do some stuff
 
-                                  /**
-                           * Records a log message to the console.
-                           */
-                  function recordLog() {
-                              console.log(/* CURSOR */ 'Recording the log')
-                          }
+                                      /**
+                               * Records a log message to the console.
+                               */
+                      function recordLog() {
+                                  console.log(/* CURSOR */ 'Recording the log')
+                              }
 
-                          recordLog()
-                      },
-                  }
-                  "
-                `)
+                              recordLog()
+                          },
+                      }"
+                    `)
             )
 
             checkDocumentCommand(
@@ -1116,203 +1083,34 @@ describe('Agent', () => {
                 obtained =>
                     expect(obtained).toMatchInlineSnapshot(
                         `
-                  "import { expect } from 'vitest'
-                  import { it } from 'vitest'
-                  import { describe } from 'vitest'
+                      "import { expect } from 'vitest'
+                      import { it } from 'vitest'
+                      import { describe } from 'vitest'
 
-                  describe('test block', () => {
-                      it('does 1', () => {
-                          expect(true).toBe(true)
-                      })
+                      describe('test block', () => {
+                          it('does 1', () => {
+                              expect(true).toBe(true)
+                          })
 
-                      it('does 2', () => {
-                          expect(true).toBe(true)
-                      })
+                          it('does 2', () => {
+                              expect(true).toBe(true)
+                          })
 
-                      it('does something else', () => {
-                          // This line will error due to incorrect usage of \`performance.now\`
-                                  /**
-                           * Returns the current time in milliseconds since the page was loaded.
-                           *
-                           * Use this to measure the duration of an operation or to profile code performance.
-                           */
-                  const startTime = performance.now(/* CURSOR */)
-                      })
-                  })
-                  "
-                `,
+                          it('does something else', () => {
+                              // This line will error due to incorrect usage of \`performance.now\`
+                                      /**
+                               * Returns the current time in milliseconds since the page was loaded.
+                               *
+                               * Use this to measure the duration of an operation or to profile code performance.
+                               */
+                      const startTime = performance.now(/* CURSOR */)
+                          })
+                      })"
+                    `,
                         explainPollyError
                     )
             )
         })
-    })
-
-    describe('Custom Commands', () => {
-        it('commands/custom, chat command, open tabs context', async () => {
-            await client.request('command/execute', {
-                command: 'cody.search.index-update',
-            })
-            // Note: The test editor has all the files opened from previous tests as open tabs,
-            // so we will need to open a new file that has not been opened before,
-            // to make sure this context type is working.
-            const trickyLogicUri = workspace.file('src', 'trickyLogic.ts')
-            await client.openFile(trickyLogicUri)
-
-            const result = (await client.request('commands/custom', {
-                key: '/countTabs',
-            })) as CustomChatCommandResult
-            expect(result.type).toBe('chat')
-            const lastMessage = await client.firstNonEmptyTranscript(result?.chatResult as string)
-            expect(trimEndOfLine(lastMessage.messages.at(-1)?.text ?? '')).toMatchInlineSnapshot(
-                `
-              "Here are the file names you have shared with me so far:
-
-              1. \`src/TestLogger.ts\`
-              2. \`src/TestClass.ts\`
-              3. \`src/sum.ts\`
-              4. \`src/squirrel.ts\`
-              5. \`src/multiple-selections.ts\`
-              6. \`src/Heading.tsx\`
-              7. \`src/example.test.ts\`
-              8. \`src/ChatColumn.tsx\`
-              9. \`src/animal.ts\`
-              10. \`src/trickyLogic.ts\`"
-            `,
-                explainPollyError
-            )
-        }, 30_000)
-
-        it('commands/custom, chat command, adds argument', async () => {
-            await client.request('command/execute', {
-                command: 'cody.search.index-update',
-            })
-            await client.openFile(animalUri)
-            const result = (await client.request('commands/custom', {
-                key: '/translate Python',
-            })) as CustomChatCommandResult
-            expect(result.type).toBe('chat')
-            const lastMessage = await client.firstNonEmptyTranscript(result?.chatResult as string)
-            expect(trimEndOfLine(lastMessage.messages.at(-1)?.text ?? '')).toMatchInlineSnapshot(
-                `
-              "Here's the Python equivalent of the provided TypeScript interface:
-
-              \`\`\`python
-              class Animal:
-                  def __init__(self, name: str, is_mammal: bool):
-                      self.name = name
-                      self.is_mammal = is_mammal
-
-                  def make_animal_sound(self) -> str:
-                      raise NotImplementedError("Subclasses must implement make_animal_sound method")
-              \`\`\`
-
-              In Python, we don't have the concept of an interface like in TypeScript, so we use a base class with abstract methods. The \`Animal\` class has the following:
-
-              1. An \`__init__\` method that takes in \`name\` (a string) and \`is_mammal\` (a boolean) as parameters and initializes the respective instance attributes.
-              2. A \`make_animal_sound\` method that is marked as an abstract method using \`raise NotImplementedError\`. This means that any concrete subclass of \`Animal\` must implement this method.
-
-              To create an instance of a specific animal, you would create a subclass of \`Animal\` and implement the \`make_animal_sound\` method. For example:
-
-              \`\`\`python
-              class Dog(Animal):
-                  def make_animal_sound(self) -> str:
-                      return "Woof!"
-
-              dog = Dog("Buddy", True)
-              print(dog.name)  # Output: Buddy
-              print(dog.is_mammal)  # Output: True
-              print(dog.make_animal_sound())  # Output: Woof!
-              \`\`\`
-
-              Note that Python doesn't have a strict type system like TypeScript, so you can omit the type annotations if you prefer."
-            `,
-                explainPollyError
-            )
-        }, 30_000)
-
-        it('commands/custom, chat command, no context', async () => {
-            await client.request('command/execute', {
-                command: 'cody.search.index-update',
-            })
-            await client.openFile(animalUri)
-            const result = (await client.request('commands/custom', {
-                key: '/none',
-            })) as CustomChatCommandResult
-            expect(result.type).toBe('chat')
-            const lastMessage = await client.firstNonEmptyTranscript(result.chatResult as string)
-            expect(trimEndOfLine(lastMessage.messages.at(-1)?.text ?? '')).toMatchInlineSnapshot(
-                `"no"`,
-                explainPollyError
-            )
-        }, 30_000)
-
-        // The context files are presented in an order in the CI that is different
-        // than the order shown in recordings when on Windows, causing it to fail.
-        it('commands/custom, chat command, current directory context', async () => {
-            await client.request('command/execute', {
-                command: 'cody.search.index-update',
-            })
-            await client.openFile(animalUri)
-            const result = (await client.request('commands/custom', {
-                key: '/countDirFiles',
-            })) as CustomChatCommandResult
-            expect(result.type).toBe('chat')
-            const lastMessage = await client.firstNonEmptyTranscript(result.chatResult as string)
-            const reply = trimEndOfLine(lastMessage.messages.at(-1)?.text ?? '')
-            expect(reply).not.includes('.cody/ignore') // file that's not located in the src/directory
-            expect(reply).toMatchInlineSnapshot(`"9"`, explainPollyError)
-        }, 30_000)
-
-        it('commands/custom, edit command, insert mode', async () => {
-            await client.request('command/execute', {
-                command: 'cody.search.index-update',
-            })
-            await client.openFile(sumUri, { removeCursor: false })
-            const result = (await client.request('commands/custom', {
-                key: '/hello',
-            })) as CustomEditCommandResult
-            expect(result.type).toBe('edit')
-            await client.taskHasReachedAppliedPhase(result.editResult as EditTask)
-
-            const originalDocument = client.workspace.getDocument(sumUri)!
-            expect(trimEndOfLine(originalDocument.getText())).toMatchInlineSnapshot(
-                `
-              "/* hello */
-              export function sum(a: number, b: number): number {
-                  /* CURSOR */
-              }
-              "
-            `,
-                explainPollyError
-            )
-        }, 30_000)
-
-        it('commands/custom, edit command, edit mode', async () => {
-            await client.request('command/execute', {
-                command: 'cody.search.index-update',
-            })
-            await client.openFile(animalUri)
-
-            const result = (await client.request('commands/custom', {
-                key: '/newField',
-            })) as CustomEditCommandResult
-            expect(result.type).toBe('edit')
-            await client.taskHasReachedAppliedPhase(result.editResult as EditTask)
-
-            const originalDocument = client.workspace.getDocument(animalUri)!
-            expect(trimEndOfLine(originalDocument.getText())).toMatchInlineSnapshot(`
-              "export interface Animal {
-                  name: string
-                  makeAnimalSound(): string
-                  isMammal: boolean
-                  logName(): void {
-                      console.log(this.name);
-                  }
-              }
-
-              "
-            `)
-        }, 30_000)
     })
 
     describe('Progress bars', () => {
@@ -1450,34 +1248,26 @@ describe('Agent', () => {
             obtained =>
                 expect(obtained).toMatchInlineSnapshot(
                     `
-              "import { expect } from 'vitest'
-              import { it } from 'vitest'
-              import { describe } from 'vitest'
+                  "import { expect } from 'vitest'
+                  import { it } from 'vitest'
+                  import { describe } from 'vitest'
 
-              /**
-               * Test block for example functionality
-               *
-               * This test block contains three test cases:
-               * - "does 1": Verifies that true is equal to true
-               * - "does 2": Verifies that true is equal to true
-               * - "does something else": Currently incomplete test case that will error due to incorrect usage of \`performance.now\`
-               */
-              describe('test block', () => {
-                  it('does 1', () => {
-                      expect(true).toBe(true)
-                  })
+                  describe('test block', () => {
+                      it('does 1', () => {
+                          expect(true).toBe(true)
+                      })
 
-                  it('does 2', () => {
-                      expect(true).toBe(true)
-                  })
+                      it('does 2', () => {
+                          expect(true).toBe(true)
+                      })
 
-                  it('does something else', () => {
-                      // This line will error due to incorrect usage of \`performance.now\`
-                      const startTime = performance.now(/* CURSOR */)
-                  })
-              })
-              "
-            `,
+                      it('does something else', () => {
+                          // This line will error due to incorrect usage of \`performance.now\`
+                                  /** Get the current high resolution time in milliseconds. */
+                  const startTime = performance.now(/* CURSOR */)
+                      })
+                  })"
+                `,
                     explainPollyError
                 )
         )
